% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bioclim.R
\name{calc_bioclim}
\alias{calc_bioclim}
\title{Calculate Bioclimatic Variables from CRU Climate Data}
\usage{
calc_bioclim(data, start_year, end_year)
}
\arguments{
\item{data}{Tidy data frame output from \code{get_cru} function. Must
contain at least CRU variables \code{"pre"}, \code{"tmn"}, and
\code{"tmx"}. User may also provide similar data in
\code{\link[terra]{SpatRaster}} format.}

\item{start_year}{Integer. Year, contained in the data, that the user wants
to start calculating bioclimatic variables from.}

\item{end_year}{Integer. Year where the user wants the calculation to stop.}
}
\value{
A \code{\link[terra]{SpatRaster}} with 19 bioclimatic variables.
}
\description{
Calculates the 19 BioClim variables from monthly climate data,
  either a data frame or SpatRasters.
}
\examples{
\dontrun{
tmin <- c(10, 12, 14, 16, 18, 20, 22, 21, 19, 17, 15, 12)
tmax <- tmin + 5
prec <- c(0, 2, 10, 30, 80, 160, 80, 20, 40, 60, 20, 0)
biovars(prec, tmin, tmax)

tmn <- tmx <- prc <- terra::rast(nrow = 1, ncol = 1, nlyr = 12)
terra::values(tmn) <- t(matrix(c(10, 12, 14, 16, 18, 20, 22, 21, 19, 17, 15, 12)))
tmx <- tmn + 5
terra::values(prc) <- t(matrix(c(0, 2, 10, 30, 80, 160, 80, 20, 40, 60, 20, 0)))
b <- biovars(prc, tmn, tmx)
as.matrix(b)
}
}
