% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cru.R
\name{get_cru}
\alias{get_cru}
\title{down}
\usage{
get_cru(
  tmp = TRUE,
  tmn = FALSE,
  tmx = FALSE,
  pre = FALSE,
  vap = FALSE,
  dtr = FALSE,
  start_year = 1901,
  end_year = 2024,
  output = "rasters"
)
}
\arguments{
\item{tmp}{Logical. Fetch  \acronym{CRU} temperature data (degrees Celsius) and return to the dataframe.
Defaults to \code{FALSE}.}

\item{tmn}{Logical. Fetch  \acronym{CRU} temperature data (degrees Celsius) and calculate monthly minimum temperature
and return to the dataframe.Defaults to \code{FALSE}.}

\item{tmx}{Logical. Fetch  \acronym{CRU} temperature data (degrees Celsius) and calculate monthly maximum temperature
and return to the dataframe. Defaults to \code{FALSE}.}

\item{pre}{Logical. Fetch  \acronym{CRU} precipitation data (milimeters/month) and return to the dataframe.}

\item{vap}{Logical. Fetch  \acronym{CRU} vapor pressure data (hectoPascals (hPa)) and return to the dataframe.}

\item{dtr}{Logical. Fetch  \acronym{CRU} diurnal temperature range (degrees Celsius) and return to the dataframe.}

\item{start_year}{Numeric. Define start year for fetched data. Defaults to \code{1901}.}

\item{end_year}{Numeric. Define end year for fetched data. Defaults to \code{2024}.}

\item{output}{Define if tabular data "tabular" or raster stack output "rasters" is the desired output.
Defaults to 'rasters'.}
}
\value{
A tidy dataframe with the requested data from the \acronym{CRU} v. 4.09 repository. If no data are requested, the user
will be prompted to set at least one argument to TRUE.
}
\description{
Access time series \acronym{TS} environmental data from the \acronym{CRU} repository.
For more information see the description of the data provided by
\acronym{CRU}, \url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
}
\examples{
\dontrun{
# Download data and create a data frame of precipitation and temperature
# without caching the data files
CRU_pre_tmp <- get_cru(pre = TRUE, tmp = TRUE)

head(CRU_pre_tmp)

library(tibble)
CRU_pre_tmp
}


}
